Fix issue #4 and remove duplicate icon
This will fix issue #4. You can read an in depth description of the
problem here:
https://github.com/mogzol/BrawlBuilder/issues/4#issuecomment-245806404

Additionaly, this will move the app icon from the designer resources
to the program's resources, so that it can be used in both the designer
and as the app icon.

Finally this makes a few small code cleanup changes.