Remove .config file, clean up buttons, bump version
This will remove the .config file that used to be required alongside
BrawlBuilder's exe file. Now instead of relying on the config file to
speicify where BrawlBuilder should look for it's dll files, it is
hardcoded into the application. I did this to clean up the base folder
and prevent confusion for users (mainly users who have file extensions
disabled).

Additionally this will change the 'Custom banner' browse button to not
inherit appearances from the groupbox, which should prevent it from
appearing transparent on some Windows versions/themes.